+++ CHANGES.rst
@@ -11,6 +11,8 @@ https://github.com/zopefoundation/Zope/blob/4.x/CHANGES.rst
 5.7.3 (unreleased)
 ------------------
 
+- Adjust the old ``App.Dialogs.MessageDialog`` to be served as HTML after
+  from PR https://github.com/zopefoundation/Zope/pull/1075 .
 
 5.7.2 (2022-12-17)
 ------------------